+++ Src/StringBuilder.fs
@@ -24,36 +24,47 @@ module AutoOpenExtensionsStringBuilder =
         // TODO: add overload with length: sb.IndexOf (c:char, from:int, length:int )
 
         /// Like list.IndexOf but for StringBuilder, returns -1 if not found
-        /// Throws ArgumentOutOfRangeException if 'from' is out of range
+        /// Throws ArgumentOutOfRangeException if 'from' is negative.
         member sb.IndexOf (c:char, from:int ) : int =
-            if from < 0 then ArgumentOutOfRangeException("from",$"StringBuilder.IndexOf: from ({from}) must be non-negative") |> raise
+            if from < 0 then
+                ArgumentOutOfRangeException("from",$"StringBuilder.IndexOf: from ({from}) must be non-negative") |> raise
             let len = sb.Length
-            if from >= len then ArgumentOutOfRangeException("from",$"StringBuilder.IndexOf: from ({from}) must be less than StringBuilder length ({len})") |> raise
+            // if from >= len then
+            //     ArgumentOutOfRangeException("from",$"StringBuilder.IndexOf: from ({from}) must be less than StringBuilder length ({len})") |> raise
             let rec find i =
-                if   i = len    then -1
-                elif sb.[i] = c then i
-                else find (i+1)
+                if i = len then
+                    -1
+                elif sb.[i] = c then
+                    i
+                else
+                    find (i+1)
             find(from)
 
         /// Like list.IndexOf but for StringBuilder, returns -1 if not found
         /// Always uses StringComparison.Ordinal
-        /// Throws ArgumentOutOfRangeException if 'from' is out of range
+        /// Throws ArgumentOutOfRangeException if 'from' is negative.
         member sb.IndexOf (t:string, from:int) : int =
             // could in theory be improved be using a rolling hash value
             // see also Array.findArray implementation
             // or https://stackoverflow.com/questions/12261344/fastest-search-method-in-stringbuilder
-            let ls = sb.Length
-            let lt = t.Length
-            if from < 0 then ArgumentOutOfRangeException("from",$"StringBuilder.IndexOf: from ({from}) must be non-negative") |> raise
-            if from >= ls then ArgumentOutOfRangeException("from",$"StringBuilder.IndexOf: from ({from}) must be less than StringBuilder length ({ls})") |> raise
-            //printfn "sb :%d t:%d" ls lt
-            let rec find ib it = // index in StringBuilder and index in search string
-                //printfn "Search at ib:%d %c for it:%d %c" ib sb.[ib] it  t.[it]
-                if  ib > ls-lt+it then -1 // not found! not enough chars left in StringBuilder to match remaining search string
-                elif sb.[ib] = t.[it]  then
-                    if it = lt-1 then ib - lt + 1 // found !
-                    else find (ib+1) (it+1)
-                else find (ib+1-it) 0
+            let lenBuilder = sb.Length
+            let lenText = t.Length
+            if from < 0 then
+                ArgumentOutOfRangeException("from",$"StringBuilder.IndexOf: from ({from}) must be non-negative") |> raise
+
+            // if from >= lenBuilder then
+            //     ArgumentOutOfRangeException("from",$"StringBuilder.IndexOf: from ({from}) must be less than StringBuilder length ({lenBuilder})") |> raise
+
+            let rec find idxBuilder idxText = // index in StringBuilder and index in search string
+                if idxBuilder > lenBuilder-lenText+idxText then
+                    -1 // not found! not enough chars left in StringBuilder to match remaining search string
+                elif sb.[idxBuilder] = t.[idxText]  then
+                    if idxText = lenText-1 then
+                        idxBuilder - lenText + 1 // found !
+                    else
+                        find (idxBuilder+1) (idxText+1)
+                else
+                    find (idxBuilder+1-idxText) 0
             find from 0
 
         /// Like list.IndexOf but for StringBuilder, returns -1 if not found